added a new test
	modified:   test/legacy_test/test_flash_attention.py
	new file:   test/legacy_test/test_scaled_dot_product_attention.py
