+++ test/legacy_test/test_flash_attention.py
@@ -76,28 +76,6 @@ def attention_naive_with_mask(q, k, v, attn_bias):
     return paddle.transpose(o, [0, 2, 1, 3])
 
 
-def attention_naive_with_bool_mask(q, k, v, bool_mask):
-    qt = paddle.transpose(q, [0, 2, 1, 3])
-    kt = paddle.transpose(k, [0, 2, 1, 3])
-    vt = paddle.transpose(v, [0, 2, 1, 3])
-
-    scale = 1.0 / np.sqrt(q.shape[-1])
-    s = paddle.matmul(qt, paddle.transpose(kt, [0, 1, 3, 2]))
-    s = paddle.scale(s, scale)
-
-    float_mask = paddle.where(
-        bool_mask,
-        paddle.to_tensor(0.0, dtype=q.dtype),
-        paddle.to_tensor(-float('inf'), dtype=q.dtype),
-    )
-
-    s = s + float_mask
-    p = F.softmax(s)
-
-    o = paddle.matmul(p, vt)
-    return paddle.transpose(o, [0, 2, 1, 3])
-
-
 is_sm80 = (
     core.is_compiled_with_cuda()
     and paddle.device.cuda.get_device_capability()[0] == 8
@@ -409,104 +387,6 @@ class TestFlashAttentionWithMaskAPI(unittest.TestCase):
         out.backward()
         out_.backward()
         np.testing.assert_allclose(out.numpy(), out_, rtol=5e-03, atol=1e-03)
-
-
-class TestAttentionWithBoolMask(unittest.TestCase):
-    def setUp(self):
-        self.place = paddle.CPUPlace()
-        self.shape = (2, 128, 8, 32)
-        self.dtype = 'float32'  # 修改为float32
-        self.dropout = 0.0
-        self.causal = False
-
-    def test_dot_scale_product_bool_mask(self):
-        # test dynamic
-        paddle.disable_static()
-
-        query = np.random.random(self.shape)
-        key = np.random.random(self.shape)
-        value = np.random.random(self.shape)
-
-        q = paddle.to_tensor(
-            query, place=self.place, dtype=self.dtype, stop_gradient=False
-        )
-        k = paddle.to_tensor(
-            key, place=self.place, dtype=self.dtype, stop_gradient=False
-        )
-        v = paddle.to_tensor(
-            value, place=self.place, dtype=self.dtype, stop_gradient=False
-        )
-
-        q_ = paddle.to_tensor(
-            query, place=self.place, dtype=self.dtype, stop_gradient=False
-        )
-        k_ = paddle.to_tensor(
-            key, place=self.place, dtype=self.dtype, stop_gradient=False
-        )
-        v_ = paddle.to_tensor(
-            value, place=self.place, dtype=self.dtype, stop_gradient=False
-        )
-
-        mask_shape = (self.shape[0], 1, self.shape[1], self.shape[1])
-        bool_mask = np.random.choice([True, False], size=mask_shape)
-
-        m = paddle.to_tensor(
-            bool_mask, place=self.place, dtype=paddle.bool, stop_gradient=False
-        )
-
-        out = scaled_dot_product_attention(
-            q, k, v, m, self.dropout, self.causal
-        )
-
-        out_ = attention_naive_with_bool_mask(q_, k_, v_, m)
-
-        out.backward()
-        out_.backward()
-
-        np.testing.assert_allclose(out.numpy(), out_, rtol=5e-03, atol=1e-03)
-
-    def test_dot_scale_product_none_mask(self):
-        # test with mask=None
-        paddle.disable_static()
-
-        query = np.random.random(self.shape)
-        key = np.random.random(self.shape)
-        value = np.random.random(self.shape)
-
-        q = paddle.to_tensor(
-            query, place=self.place, dtype=self.dtype, stop_gradient=False
-        )
-        k = paddle.to_tensor(
-            key, place=self.place, dtype=self.dtype, stop_gradient=False
-        )
-        v = paddle.to_tensor(
-            value, place=self.place, dtype=self.dtype, stop_gradient=False
-        )
-
-        q_ = paddle.to_tensor(
-            query, place=self.place, dtype=self.dtype, stop_gradient=False
-        )
-        k_ = paddle.to_tensor(
-            key, place=self.place, dtype=self.dtype, stop_gradient=False
-        )
-        v_ = paddle.to_tensor(
-            value, place=self.place, dtype=self.dtype, stop_gradient=False
-        )
-
-        m = None
-
-        out = scaled_dot_product_attention(
-            q, k, v, m, self.dropout, self.causal
-        )
-
-        out_ = attention_naive(q_, k_, v_, self.causal)
-
-        out.backward()
-        out_.backward()
-
-        np.testing.assert_allclose(
-            out.numpy(), out_.numpy(), rtol=5e-03, atol=1e-03
-        )
 
 
 class TestFlashAttentionAPITest1(TestFlashAttentionAPI):
